refactor(structure) creacion del las carpetas cliente y servidor con respectivas carpetas mvc
