+++ specs/data/cliente.properties
@@ -1,0 +1,2 @@
+usuario.name=Jhon
+usuario.contrase\u00f1a=1234
