+++ specs/data/servidor.properties
@@ -1,0 +1,31 @@
+#lista de usuarios
+cantidadUsuarios=8
+
+usuario1.name=Jhon
+usuario1.contrase\u00f1a=1234
+
+usuario2.name=Maria
+usuario2.contrase\u00f1a=123456
+
+usuario3.name=Pedro
+usuario3.contrase\u00f1a=abc123
+
+usuario4.name=Lucia
+usuario4.contrase\u00f1a=pass4321
+
+usuario5.name=Carlos
+usuario5.contrase\u00f1a=car2025
+
+usuario6.name=Ana
+usuario6.contrase\u00f1a=ana321
+
+usuario7.name=Luis
+usuario7.contrase\u00f1a=luis987
+
+usuario8.name=Elena
+usuario8.contrase\u00f1a=elena555
+
+#base de datos
+db.url=jdbc:mysql://localhost:3308/segundoparcial
+db.user=root
+db.password=
