+++ src/udistrital/avanzada/pacman/cliente/Control/ControlVentana.java
@@ -1,0 +1,15 @@
+package udistrital.avanzada.pacman.cliente.control;
+
+/**
+ * Clase ControlVentana.
+ * <p>
+ * Descripción:
+ * </p>
+ *
+ * @author Mauricio
+ * @version 1.0
+ * @since 2025-11-05
+ */
+public class ControlVentana {
+
+}
