+++ src/udistrital/avanzada/pacman/cliente/Modelo/Conexion/ConexionProperties.java
@@ -1,4 +1,4 @@
-package udistrital.avanzada.pacman.Modelo.ModeloConexion;
+package udistrital.avanzada.pacman.cliente.modelo.Conexion;
 
 /**
  * Clase ConexionProperties.
@@ -6,7 +6,7 @@ package udistrital.avanzada.pacman.Modelo.ModeloConexion;
  * Descripción:
  * </p>
  *
- * @author Diego
+ * @author Mauricio
  * @version 1.0
  * @since 2025-11-05
  */
