+++ src/udistrital/avanzada/pacman/cliente/Modelo/DAO/PropertiesDAO.java
@@ -1,0 +1,15 @@
+package udistrital.avanzada.pacman.cliente.modelo.DAO;
+
+/**
+ * Clase PropertiesDAO.
+ * <p>
+ * Descripción:
+ * </p>
+ *
+ * @author Mauricio
+ * @version 1.0
+ * @since 2025-11-05
+ */
+public class PropertiesDAO {
+
+}
