+++ src/udistrital/avanzada/pacman/cliente/Vista/VentanaPrincipal.java
@@ -1,0 +1,15 @@
+package udistrital.avanzada.pacman.cliente.vista;
+
+/**
+ * Clase VentanaPrincipal.
+ * <p>
+ * Descripción:
+ * </p>
+ *
+ * @author Diego
+ * @version 1.0
+ * @since 2025-11-05
+ */
+public class VentanaPrincipal {
+
+}
