+++ src/udistrital/avanzada/pacman/servidor/Modelo/ModeloConexion/ConexionProperties.java
@@ -1,0 +1,15 @@
+package udistrital.avanzada.pacman.servidor.Modelo.ModeloConexion;
+
+/**
+ * Clase ConexionProperties.
+ * <p>
+ * Descripción:
+ * </p>
+ *
+ * @author Diego
+ * @version 1.0
+ * @since 2025-11-05
+ */
+public class ConexionProperties {
+
+}
