+++ src/udistrital/avanzada/pacman/servidor/Modelo/ModeloConexion/IConexionProperties.java
@@ -1,0 +1,15 @@
+package udistrital.avanzada.pacman.servidor.Modelo.ModeloConexion;
+
+/**
+ * Clase IConexionProperties.
+ * <p>
+ * Descripción:
+ * </p>
+ *
+ * @author Diego
+ * @version 1.0
+ * @since 2025-11-05
+ */
+public interface IConexionProperties {
+
+}
