+++ src/udistrital/avanzada/pacman/servidor/Vista/VentanaJuego.java
@@ -1,0 +1,19 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package udistrital.avanzada.pacman.servidor.Vista;
+
+/**
+ * Clase VentanaJuego.
+ * <p>
+ * Descripción: ventana que tiene cada jugador con su propio juego
+ * </p>
+ *
+ * @author Mauricio
+ * @version 1.0
+ * @since 2025-11-05
+ */
+public class VentanaJuego {
+
+}
